+++ nuriggun/asgi.py
@@ -1,12 +1,3 @@
-"""
-ASGI config for nuriggun project.
-
-It exposes the ASGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.2/howto/deployment/asgi/
-"""
-
 import os
 
 from django.core.asgi import get_asgi_application
