+++ requirements.txt
@@ -17,6 +17,7 @@ django-cors-headers==4.0.0
 django-restframework==0.0.1
 djangorestframework==3.14.0
 djangorestframework-simplejwt==5.2.2
+Faker==18.11.2
 frozenlist==1.3.3
 idna==3.4
 multidict==6.0.4
@@ -30,6 +31,7 @@ protobuf==3.20.3
 pycparser==2.21
 PyJWT==2.7.0
 PyMySQL==1.0.3
+python-dateutil==2.8.2
 python-dotenv==1.0.0
 python-env==1.0.0
 python3-openid==3.2.0
